+++ CHANGELOG.md
@@ -16,6 +16,14 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2026-02-09
+
+- richat-client-v8.1.0
+
+### Features
+
+- client: geyser access from client ([#189](https://github.com/lamports-dev/richat/pull/189))
+
 ## 2026-01-25
 
 - richat-filter-v8.1.0